+++ examples/configuration/app.js
@@ -1,22 +1,26 @@
+/*!
+ *  Example configuration autoloader.
+ *
+ *  @author Jarrad Seers <jarrad@jarradseers.com>
+ *  @created 09/08/2012 NZST
+ */
+
+/**
+ *  Module dependencies.
+ */
+
 var express = require('express')
-	,	load = require('../../');
+  , load = require('../../');
 
 var app = express();
 
-load('configs', app);
-
-for (var config in app.configs) {
-	app.configure(config, function() {
-		for (var key in app.configs[config]) {
-			app.set(key, app.configs[config][key]);
-		}
-	});
-}
+load('configs', app)
+  .autoConfigure(app.configs);
 
 app.listen(app.settings.port);
 
 console.log('%s running in %s mode on port %s' 
-	, app.settings.title 
-	, app.settings.env
+  , app.settings.title 
+  , app.settings.env
   , app.settings.port
 );